pyproject.toml
[project]
name = "handover-toy-dataset-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pathfinding>=1.0.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
]
